refactor: split Reporter file
