+++ lib/TestRunners/Reporter.js
@@ -1,5 +1,7 @@
 const colors = require('colors/safe');
 const { createCoverageMap } = require('istanbul-lib-coverage');
+const { Test } = require('./Test');
+const { Suite } = require('./Suite');
 
 const symbols = {
     ok: '✓',
@@ -238,196 +240,63 @@ ${colors.red(`${symbols.err} ${report.failed.length} failed`)}
 ${colors.grey(`${symbols.bang} ${report.skipped.length} skipped`)}`;
 }
 
-class Test {
-    get path() {
-        let result = [];
-        let parent = this;
-        while (parent && !parent.isRoot) {
-            result.unshift(parent.title);
-            parent = parent.parent;
-        }
-
-        return result;
-    }
-
-    constructor(title) {
-        this.title = title;
-        this.passes = 0;
-        this.failes = 0;
-    }
-
-    isPassed() {
-        return (this.passes > 0 && this.failes === 0);
-    }
-
-    isFailed() {
-        return this.failes > 0;
-    }
-
-    pass(howMany = 1) {
-        this.passes += howMany;
-    }
-
-    fail(error, howMany = 1) {
-        this.failes += howMany;
-
-        stringifyDiffObjs(error);
-        // more than one error per test
-        if (this.error && error instanceof Error) {
-            if (error.multiple) {
-                this.error.multiple = [
-                    ...(this.error.multiple || []),
-                    ...error.multiple,
-                ];
-            } else {
-                this.error.multiple = [
-                    ...(this.error.multiple || []),
-                    error,
-                ];
-            }
-        } else {
-            this.error = error;
-        }
-    }
+function isEmptyCoverage(coverage) {
+    let summary = coverage.getCoverageSummary();
+    return !summary.data ||
+        summary.data.lines.pct === 'Unknown' ||
+        summary.data.statements.pct === 'Unknown' ||
+        summary.data.functions.pct === 'Unknown' ||
+        summary.data.branches.pct === 'Unknown';
 }
 
-class Suite {
-    get isRoot() {
-        return !this.parent;
+function formatCoverage(coverage) {
+    let summary = coverage.getCoverageSummary();
+    let output = `${colors.bold('Coverage')}\n`;
+    output += `${printLine(summary, 'statements')}\n`;
+    output += `${printLine(summary, 'branches')}\n`;
+    output += `${printLine(summary, 'functions')}\n`;
+    output += printLine(summary, 'lines');
+
+    return output;
+}
+
+function formatReport(report) {
+    let output = report.suite.suites.map((suite) => printSuite(suite)).join('\n');
+    if (output) {
+        output += '\n';
     }
 
-    get path() {
-        if (this.isRoot) {
-            return [];
-        }
-
-        let result = [];
-        let parent = this;
-        while (parent) {
-            result.unshift(parent.title);
-            parent = parent.parent;
-        }
-
-        return result;
+    output += report.suite.tests.map((test) => printTest(test)).join('\n');
+    if (output) {
+        output += '\n';
     }
 
-    constructor(title = '') {
-        this.title = title;
-        this.suites = [];
-        this.tests = [];
+    if (report.failed && report.failed.length) {
+        output += '\n';
+        output += printFailed(report.failed);
+        output += '\n';
     }
 
-    find(path, where = 'tests') {
-        if (!path.length) {
-            return this;
-        }
+    if (output) {
+        output += '\n';
+    }
+    output += printSummary(report);
 
-        path = path.slice(0);
-        let title = path[path.length - 1];
-        let parent = this.find(path.slice(0, -1), 'suites');
-        if (!parent) {
-            return null;
-        }
-
-        return parent[where].find((child) => child.title === title) || null;
+    if (output) {
+        output += '\n';
     }
 
-    ensureSuite(path) {
-        if (!path.length) {
-            return this;
-        }
-
-        path = path.slice(0);
-        let title = path[path.length - 1];
-        let parent = this.ensureSuite(path.slice(0, -1));
-        let child = parent.suites.find((child) => child.title === title);
-        if (!child) {
-            child = new Suite(title);
-            child.parent = parent;
-            parent.suites.push(child);
-        }
-
-        return child;
+    let coverageMap = createCoverageMap(report.coverage);
+    if (!isEmptyCoverage(coverageMap)) {
+        output += '\n';
+        output += formatCoverage(coverageMap);
+        output += '\n';
     }
 
-    addTest(test, path) {
-        if (typeof test === 'string') {
-            test = new Test(test);
-        } else if (Array.isArray(test)) {
-            path = test;
-            test = new Test(path[path.length - 1]);
-        }
-
-        path = path ? path : test.path;
-        let parent = this.ensureSuite(path.slice(0, -1));
-        let child = parent.tests.find((child) => child.title === test.title);
-        if (!child) {
-            child = new Test(test.title);
-            child.parent = parent;
-            parent.tests.push(child);
-        }
-
-        return child;
-    }
+    return output;
 }
 
 class Reporter {
-    static formatReport(report) {
-        let output = report.suite.suites.map((suite) => printSuite(suite)).join('\n');
-        if (output) {
-            output += '\n';
-        }
-
-        output += report.suite.tests.map((test) => printTest(test)).join('\n');
-        if (output) {
-            output += '\n';
-        }
-
-        if (report.failed && report.failed.length) {
-            output += '\n';
-            output += printFailed(report.failed);
-            output += '\n';
-        }
-
-        if (output) {
-            output += '\n';
-        }
-        output += printSummary(report);
-
-        if (output) {
-            output += '\n';
-        }
-
-        let coverageMap = createCoverageMap(report.coverage);
-        if (!this.isEmptyCoverage(coverageMap)) {
-            output += '\n';
-            output += this.formatCoverage(coverageMap);
-            output += '\n';
-        }
-
-        return output;
-    }
-
-    static formatCoverage(coverage) {
-        let summary = coverage.getCoverageSummary();
-        let output = `${colors.bold('Coverage')}\n`;
-        output += `${printLine(summary, 'statements')}\n`;
-        output += `${printLine(summary, 'branches')}\n`;
-        output += `${printLine(summary, 'functions')}\n`;
-        output += printLine(summary, 'lines');
-
-        return output;
-    }
-
-    static isEmptyCoverage(coverage) {
-        let summary = coverage.getCoverageSummary();
-        return !summary.data ||
-            summary.data.lines.pct === 'Unknown' ||
-            summary.data.statements.pct === 'Unknown' ||
-            summary.data.functions.pct === 'Unknown' ||
-            summary.data.branches.pct === 'Unknown';
-    }
-
     constructor(name) {
         this.name = name;
         this.suite = new Suite();
@@ -438,14 +307,14 @@ class Reporter {
     }
 
     add(test) {
-        return this.suite.addTest(test);
+        return this.suite.add(test);
     }
 
     skip(test) {
         if (test instanceof Test) {
-            test = this.suite.find(test.path) || this.suite.addTest(test);
+            test = this.suite.find(test.path) || this.suite.add(test);
         } else {
-            test = this.suite.find(test) || this.suite.addTest(test);
+            test = this.suite.find(test) || this.suite.add(test);
         }
 
         this.skipped.push(test);
@@ -453,9 +322,9 @@ class Reporter {
 
     pass(test) {
         if (test instanceof Test) {
-            test = this.suite.find(test.path) || this.suite.addTest(test);
+            test = this.suite.find(test.path) || this.suite.add(test);
         } else {
-            test = this.suite.find(test) || this.suite.addTest(test);
+            test = this.suite.find(test) || this.suite.add(test);
         }
 
         test.pass();
@@ -464,9 +333,9 @@ class Reporter {
 
     fail(test, error) {
         if (test instanceof Test) {
-            test = this.suite.find(test.path) || this.suite.addTest(test);
+            test = this.suite.find(test.path) || this.suite.add(test);
         } else {
-            test = this.suite.find(test) || this.suite.addTest(test);
+            test = this.suite.find(test) || this.suite.add(test);
         }
 
         test.fail(error);
@@ -504,6 +373,5 @@ class Reporter {
     }
 }
 
-module.exports.Test = Test;
-module.exports.Suite = Suite;
+module.exports.formatReport = formatReport;
 module.exports.Reporter = Reporter;
